requirements.txt
fastapi
uvicorn
sse-starlette
pydantic
pydantic-settings
httpx
python-dotenv
loguru
async-lru
google-generativeai
